vochain/indexer: use sqlite engine in sqlc
It's complete enough that it can parse all of our queries,
so we no longer need to (ab)use the mysql engine and hope for the best.
We update to the latest sqlc master to use a version that works well.

It even solves the issue we were having with LIMIT/OFFSET and args,
since that issue was particular to the mysql parser.

One minor change is that limit/offset parameters are now int64,
which is larger than the old int32. This seems harmless.

Another minor change is that more query parameters are interface{} now.
That seems mildly unfortunate, but it's ultimately also harmless.
If we fed the wrong types into the queries, our tests would blow up.

+++ vochain/indexer/indexer.go
@@ -30,7 +30,7 @@ import (
 	_ "github.com/mattn/go-sqlite3"
 )
 
-//go:generate go run github.com/kyleconroy/sqlc/cmd/sqlc@v1.17.2 generate
+//go:generate go run github.com/kyleconroy/sqlc/cmd/sqlc@734e06ede7e68dc76e53f41727285abe5301dc69 generate
 
 //go:embed migrations/*.sql
 var embedMigrations embed.FS
@@ -564,8 +564,8 @@ func (idx *Indexer) OnTransferTokens(tx *vochaintx.TokenTransfer) {
 func (idx *Indexer) GetTokenTransfersByFromAccount(from []byte, offset, maxItems int32) ([]*indexertypes.TokenTransferMeta, error) {
 	ttFromDB, err := idx.readOnlyQuery.GetTokenTransfersByFromAccount(context.TODO(), indexerdb.GetTokenTransfersByFromAccountParams{
 		FromAccount: from,
-		Limit:       maxItems,
-		Offset:      offset,
+		Limit:       int64(maxItems),
+		Offset:      int64(offset),
 	})
 	if err != nil {
 		return nil, err
